+++ themebuilder-scss/tests/data/scss/widgets/generic/_index.scss
@@ -1,9 +1,3 @@
-$color: null !default;
-$size: null !default;
-
-@use "colors" with ($color: $color);
-@use "sizes" with ($size: $size);
-
 // public widgets
 @use "./accordion";
 @use "./dateBox"
